feat(admin-cb): add PaddingSelect and move fields to separate folder
