+++ src/admin/content-block/components/fields/ColorSelect/ColorSelect.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/settings/colors';
-@import '../../../../styles/settings/variables';
+@import '../../../../../styles/settings/colors';
+@import '../../../../../styles/settings/variables';
 
 .c-color-select {
 	.react-select__control {
