+++ src/admin/content-block/components/fields/index.ts
@@ -1,0 +1,3 @@
+export { default as AlignSelect } from './AlignSelect/AlignSelect';
+export { default as ColorSelect } from './ColorSelect/ColorSelect';
+export { default as PaddingSelect } from './PaddingSelect/PaddingSelect';
