+++ src/admin/content-block/components/index.ts
@@ -1,4 +1,4 @@
-export { default as AlignSelect } from './AlignSelect/AlignSelect';
-export { default as ColorSelect } from './ColorSelect/ColorSelect';
+export * from './fields';
 export { default as ContentBlockForm } from './ContentBlockForm/ContentBlockForm';
 export { default as ContentBlockPreview } from './ContentBlockPreview/ContentBlockPreview';
+export * from './wrappers';
